git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@2166 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/io/StringInputStream.h
@@ -43,7 +43,7 @@ namespace org {
         long long int available(bool blocked=false);
       private:
         std::string _data;
-        long long int _position;
+        unsigned int _position;
       };
     }
   }
